+++ src/regexs.js
@@ -51,11 +51,56 @@ const GITHUB_REGEX_STR = '[a-z\\d](?:[a-z\\d]|-(?=[a-z\\d])){0,38}';
 export const TWITTER_REGEX = /^@[a-z0-9_]{1,15}$/i;
 export const GITHUB_REGEX = new RegExp(`^${GITHUB_REGEX_STR}$`, 'i');
 
-// @TODO: Test this regex
-// eslint-disable-next-line
-// /^(?:(?:https?|ftp):\/\/)(?:\S+(?::\S*)?@)?(?:(?!10(?:\.\d{1,3}){3})(?!127(?:\.\d{1,3}){3})(?!169\.254(?:\.\d{1,3}){2})(?!192\.168(?:\.\d{1,3}){2})(?!172\.(?:1[6-9]|2\d|3[0-1])(?:\.\d{1,3}){2})(?:[1-9]\d?|1\d\d|2[01]\d|22[0-3])(?:\.(?:1?\d{1,2}|2[0-4]\d|25[0-5])){2}(?:\.(?:[1-9]\d?|1\d\d|2[0-4]\d|25[0-4]))|(?:(?:[a-z\u00a1-\uffff0-9]+-?)*[a-z\u00a1-\uffff0-9]+)(?:\.(?:[a-z\u00a1-\uffff0-9]+-?)*[a-z\u00a1-\uffff0-9]+)*(?:\.(?:[a-z\u00a1-\uffff]{2,})))(?::\d{2,5})?(?:\/[^\s]*)?$/i
-// And if it works, use it instead
+// @TODO: Discontinue usage of this regexps, in favor of HTTP_URL_REGEX
 export const URL_REGEX = /^(http:\/\/|https:\/\/)/i;
+
+// Inspired by https://gist.github.com/dperini/729294, but doesn't match FTP URLs
+/* eslint-disable */
+export const HTTP_URL_REGEX = new RegExp(
+  '^' +
+    // protocol identifier (optional)
+    // short syntax // still required
+    // NOTE: We removed "|ftp"
+    '(?:(?:(?:https?):)?\\/\\/)' +
+    // user:pass BasicAuth (optional)
+    '(?:\\S+(?::\\S*)?@)?' +
+    '(?:' +
+      // IP address exclusion
+      // private & local networks
+      '(?!(?:10|127)(?:\\.\\d{1,3}){3})' +
+      '(?!(?:169\\.254|192\\.168)(?:\\.\\d{1,3}){2})' +
+      '(?!172\\.(?:1[6-9]|2\\d|3[0-1])(?:\\.\\d{1,3}){2})' +
+      // IP address dotted notation octets
+      // excludes loopback network 0.0.0.0
+      // excludes reserved space >= 224.0.0.0
+      // excludes network & broadcast addresses
+      // (first & last IP address of each class)
+      '(?:[1-9]\\d?|1\\d\\d|2[01]\\d|22[0-3])' +
+      '(?:\\.(?:1?\\d{1,2}|2[0-4]\\d|25[0-5])){2}' +
+      '(?:\\.(?:[1-9]\\d?|1\\d\\d|2[0-4]\\d|25[0-4]))' +
+    '|' +
+      // host & domain names, may end with dot
+      // can be replaced by a shortest alternative
+      // (?![-_])(?:[-\\w\\u00a1-\\uffff]{0,63}[^-_]\\.)+
+      '(?:' +
+        '(?:' +
+          '[a-z0-9\\u00a1-\\uffff]' +
+          '[a-z0-9\\u00a1-\\uffff_-]{0,62}' +
+        ')?' +
+        '[a-z0-9\\u00a1-\\uffff]\\.' +
+      ')+' +
+      // TLD identifier name, may end with dot
+      // NOTE: "|xn--[a-z0-9]+" is our addition to support IDNs like "http://xn--80aaxitdbjk.xn--p1ai",
+      // they can be used in a browser, so we consider them valid
+      '(?:[a-z\\u00a1-\\uffff]{2,}\\.?|xn--[a-z0-9]+)' +
+    ')' +
+    // port number (optional)
+    '(?::\\d{2,5})?' +
+    // resource path (optional)
+    '(?:[/?#]\\S*)?' +
+  '$', 'i',
+);
+/* eslint-enable */
 
 // E.g. https://gist.github.com/jancurn/2dbe83fea77c439b1119fb3f118513e7
 export const GITHUB_GIST_URL_REGEX = new RegExp(`^https:\\/\\/gist\\.github\\.com\\/${GITHUB_REGEX_STR}\\/[0-9a-f]{32}$`, 'i');