+++ test/utilities.js
@@ -185,3 +185,143 @@ describe('utilities', () => {
         expect(nicePath).to.be.eql('12345-api-example-crawler');
     });
 });
+
+describe('timeoutPromise()', () => {
+    it('should work when promise resolves in time', async () => {
+        const promise = new Promise((resolve) => {
+            setTimeout(() => resolve('xxx'), 100);
+        });
+        const timeoutPromise = utils.timeoutPromise(promise, 200);
+        expect(await timeoutPromise).to.be.eql('xxx');
+    });
+
+    it('should work when promise rejects in time', async () => {
+        const error = new Error('some-fail');
+        const promise = new Promise((resolve, reject) => {
+            setTimeout(() => reject(error), 100);
+        });
+        const timeoutPromise = utils.timeoutPromise(promise, 200);
+
+        try {
+            await timeoutPromise;
+            throw new Error('This should have failed!');
+        } catch (err) {
+            expect(err.message).to.be.eql('some-fail');
+        }
+    });
+
+    it('should work when promise timeouts in time', async () => {
+        const error = new Error('some-fail');
+        const promise = new Promise((resolve, reject) => {
+            setTimeout(() => reject(error), 200);
+        });
+        const timeoutPromise = utils.timeoutPromise(promise, 100);
+
+        try {
+            await timeoutPromise;
+            throw new Error('This should have failed!');
+        } catch (err) {
+            expect(err.message).to.be.eql('Promise has timed-out');
+        }
+    });
+
+    it('should support custom error message', async () => {
+        const error = new Error('some-fail');
+        const promise = new Promise((resolve, reject) => {
+            setTimeout(() => reject(error), 200);
+        });
+        const timeoutPromise = utils.timeoutPromise(promise, 100, 'Custom error message');
+
+        try {
+            await timeoutPromise;
+            throw new Error('This should have failed!');
+        } catch (err) {
+            expect(err.message).to.be.eql('Custom error message');
+        }
+    });
+
+
+    describe('#makeInputJsFieldsReadable()', () => {
+        it('should correctly handle normal functions, arrow functions and JS code', () => {
+            /* eslint-disable */
+            const json = `{
+                "cookiesPersistence": "PER_PROCESS",
+                "disableWebSecurity": true,
+                "arrowFunction": "async (a, b) => a + b; ",
+                "loadCss": false,
+                "normalFunction": "function pageFunction(context) {\\n    // called on every page the crawler visits, use it to extract data from it\\n    const $ = context.jQuery;\\n    return 'xxxx';\\n}",
+                "proxyConfiguration": {
+                    "useApifyProxy": false
+                },
+                "rotateUserAgents": false,
+                "someCode": "const a = 5;\\nfunction sum (a, b) {\\n    return a + b;\\n}\\nsum(a, 10);"
+            }`;
+            /* eslint-enable */
+
+            const given = utils.makeInputJsFieldsReadable(json, ['normalFunction', 'arrowFunction', 'someCode'], 4);
+            const expected = `{
+    "cookiesPersistence": "PER_PROCESS",
+    "disableWebSecurity": true,
+    "arrowFunction": async (a, b) => a + b,
+    "loadCss": false,
+    "normalFunction": function pageFunction(context) {
+        // called on every page the crawler visits, use it to extract data from it
+        const $ = context.jQuery;
+        return 'xxxx';
+    },
+    "proxyConfiguration": {
+        "useApifyProxy": false
+    },
+    "rotateUserAgents": false,
+    "someCode": \`const a = 5;
+        function sum (a, b) {
+            return a + b;
+        }
+        sum(a, 10);\`
+}`;
+
+            expect(given).to.be.eql(expected);
+        });
+
+        it('should not fail on invalid JS code', () => {
+            /* eslint-disable */
+            const json = `{
+                "cookiesPersistence": "PER_PROCESS",
+                "arrowFunction": "async (a, b) => a + b; ",
+                "loadCss": false,
+                "normalFunction": "function INVALID!!! pageFunction(context) {\\n    // called on every page the crawler visits, use it to extract data from it\\n    const $ = context.jQuery;\\n    return 'xxxx';\\n}"
+            }`;
+            /* eslint-enable */
+
+            const given = utils.makeInputJsFieldsReadable(json, ['normalFunction', 'arrowFunction', 'someCode'], 4);
+            /* eslint-disable */
+            const expected = `{
+    "cookiesPersistence": "PER_PROCESS",
+    "arrowFunction": async (a, b) => a + b,
+    "loadCss": false,
+    "normalFunction": "function INVALID!!! pageFunction(context) {\\n    // called on every page the crawler visits, use it to extract data from it\\n    const $ = context.jQuery;\\n    return 'xxxx';\\n}"
+}`;
+            /* eslint-enable */
+
+            expect(given).to.be.eql(expected);
+        });
+
+
+        it('should support global spaces', () => {
+            const json = `{
+                "cookiesPersistence": "PER_PROCESS",
+                "arrowFunction": "async (a, b) => a + b; ",
+                "loadCss": false
+            }`;
+
+            const given = utils.makeInputJsFieldsReadable(json, ['normalFunction', 'arrowFunction', 'someCode'], 4, 4);
+            const expected = `{
+        "cookiesPersistence": "PER_PROCESS",
+        "arrowFunction": async (a, b) => a + b,
+        "loadCss": false
+    }`;
+
+            expect(given).to.be.eql(expected);
+        });
+    });
+});